refactor: register cameraApp module
