+++ resources/js/app.js
@@ -2,6 +2,8 @@ import './bootstrap';
 
 import Alpine from 'alpinejs';
 
+import './cameraApp';
+
 window.Alpine = Alpine;
 
 Alpine.start();
